backend/pyproject.toml
[tool.poetry]
name = "fastapi-mongo-starter-backend"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.99.1"
uvicorn = "^0.23.2"
pydantic = "^1.10.12"
requests = "^2.31.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
fastapi-users = {extras = ["oauth", "beanie"], version = "^12.1.2"}
motor = "^3.3.2"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
autoflake = "^2.3.0"
pytest = "^7.4.3"
ipython = "^8.16.1"
pytest-cov = "^4.1.0"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.cruft]
skip = [".env"]

[tool.pytest.ini_options]
asyncio_mode = 'auto'
filterwarnings = [
    "ignore::DeprecationWarning"
]

[tool.black]
line-length = 120
